+++ packages/test-utils/src/module.d.ts
@@ -1,4 +1,4 @@
 declare module 'fela-sort-media-query-mobile-first' {
-  import {TPlugin} from 'packages/test-utils/src/index'
-  export default function (): TPlugin
+  import {TEnhancer} from 'fela'
+  export default function (): TEnhancer
 }